MAR-3063. Refactoring slices for Ebooks pages. Added new fields animation and containerSize for all slices using in Ebooks pages. Added props checks.
