+++ .slicemachine/assets/client/prismicSlices/pageParts/CardsWithIcon/mocks.json
@@ -12,16 +12,16 @@
           },
           "alt": "Placeholder image",
           "copyright": null,
-          "url": "https://images.unsplash.com/photo-1544731612-de7f96afe55f?w=900&h=500&fit=crop"
+          "url": "https://images.unsplash.com/photo-1515378960530-7c0da6231fb1?w=900&h=500&fit=crop"
         },
         "title": [
           {
             "type": "heading1",
-            "text": "Integrate proactive metrics",
+            "text": "Deliver strategic deliverables",
             "spans": []
           }
         ],
-        "description": "envisioneer interactive synergies"
+        "description": "grow front-end schemas"
       },
       {
         "iconImage": {
@@ -31,18 +31,78 @@
           },
           "alt": "Placeholder image",
           "copyright": null,
-          "url": "https://images.unsplash.com/photo-1576662712957-9c79ae1280f8?w=900&h=500&fit=crop"
+          "url": "https://images.unsplash.com/photo-1515378960530-7c0da6231fb1?w=900&h=500&fit=crop"
         },
         "title": [
           {
             "type": "heading1",
-            "text": "Engage B2B platforms",
+            "text": "Seize best-of-breed e-commerce",
             "spans": []
           }
         ],
-        "description": "cultivate real-time e-commerce"
+        "description": "brand sexy markets"
+      },
+      {
+        "iconImage": {
+          "dimensions": {
+            "width": 900,
+            "height": 500
+          },
+          "alt": "Placeholder image",
+          "copyright": null,
+          "url": "https://images.unsplash.com/photo-1544731612-de7f96afe55f?w=900&h=500&fit=crop"
+        },
+        "title": [
+          {
+            "type": "heading1",
+            "text": "Extend viral web-readiness",
+            "spans": []
+          }
+        ],
+        "description": "integrate dynamic blockchains"
+      },
+      {
+        "iconImage": {
+          "dimensions": {
+            "width": 900,
+            "height": 500
+          },
+          "alt": "Placeholder image",
+          "copyright": null,
+          "url": "https://images.unsplash.com/photo-1493119508027-2b584f234d6c?w=900&h=500&fit=crop"
+        },
+        "title": [
+          {
+            "type": "heading1",
+            "text": "Transform interactive niches",
+            "spans": []
+          }
+        ],
+        "description": "extend ubiquitous networks"
+      },
+      {
+        "iconImage": {
+          "dimensions": {
+            "width": 900,
+            "height": 500
+          },
+          "alt": "Placeholder image",
+          "copyright": null,
+          "url": "https://images.unsplash.com/photo-1587905069134-008460d7a636?w=900&h=500&fit=crop"
+        },
+        "title": [
+          {
+            "type": "heading1",
+            "text": "Aggregate next-generation vortals",
+            "spans": []
+          }
+        ],
+        "description": "innovate customized portals"
       }
     ],
-    "primary": {}
+    "primary": {
+      "animation": "zoom-in",
+      "containerSize": "1240"
+    }
   }
 ]
