+++ .slicemachine/assets/client/prismicSlices/pageParts/FormSlice/mocks.json
@@ -8,10 +8,12 @@
       "title": [
         {
           "type": "heading1",
-          "text": "Grow end-to-end architectures",
+          "text": "Incubate next-generation architectures",
           "spans": []
         }
       ],
+      "containerSize": "1026",
+      "animation": "zoom-in",
       "image": {
         "dimensions": {
           "width": 900,
@@ -19,10 +21,10 @@
         },
         "alt": "Placeholder image",
         "copyright": null,
-        "url": "https://images.unsplash.com/photo-1491975474562-1f4e30bc9468?w=900&h=500&fit=crop"
+        "url": "https://images.unsplash.com/photo-1591012911207-0dbac31f37da?w=900&h=500&fit=crop"
       },
-      "bookPath": "strategize integrated content",
-      "bookName": "generate open-source solutions"
+      "bookPath": "extend dot-com methodologies",
+      "bookName": "engineer best-of-breed interfaces"
     }
   }
 ]
