+++ .slicemachine/assets/client/prismicSlices/pageParts/PostsSectionSlice/index.stories.js
@@ -18,7 +18,7 @@ export const _DefaultSlice = () => ({
   },
   data() {
     return {
-      mock: {"variation":"default-slice","name":"Default slice","slice_type":"posts_section_slice","items":[{"theme":"light","data":{"link_type":"Web","url":"http://google.com"}},{"theme":"dark","data":{"link_type":"Web","url":"https://prismic.io"}},{"theme":"light","data":{"link_type":"Web","url":"http://twitter.com"}},{"theme":"light","data":{"link_type":"Web","url":"https://prismic.io"}},{"theme":"dark","data":{"link_type":"Web","url":"https://prismic.io"}}],"primary":{"animation":"fade-up"},"id":"_DefaultSlice"}
+      mock: {"variation":"default-slice","name":"Default slice","slice_type":"posts_section_slice","items":[{"theme":"light","data":{"link_type":"Web","url":"https://slicemachine.dev"}},{"theme":"light","data":{"link_type":"Web","url":"https://prismic.io"}},{"theme":"light","data":{"link_type":"Web","url":"https://prismic.io"}},{"theme":"light","data":{"link_type":"Web","url":"http://twitter.com"}},{"theme":"light","data":{"link_type":"Web","url":"http://twitter.com"}}],"primary":{"animation":"fade-up"},"id":"_DefaultSlice"}
     }
   },
   template: '<SliceZone :slices="[mock]" :resolver="resolve" />'
@@ -37,7 +37,7 @@ export const _EbookPostsSectionSlice = () => ({
   },
   data() {
     return {
-      mock: {"variation":"ebookPostsSectionSlice","name":"EbookPostsSectionSlice","slice_type":"posts_section_slice","items":[{"theme":"light","data":{"link_type":"Web","url":"https://slicemachine.dev"}},{"theme":"light","data":{"link_type":"Web","url":"https://slicemachine.dev"}},{"theme":"light","data":{"link_type":"Web","url":"https://prismic.io"}},{"theme":"light","data":{"link_type":"Web","url":"https://slicemachine.dev"}},{"theme":"light","data":{"link_type":"Web","url":"http://google.com"}}],"primary":{"animation":"fade-up","title":[{"type":"heading1","text":"Brand sticky functionalities","spans":[]}]},"id":"_EbookPostsSectionSlice"}
+      mock: {"variation":"ebookPostsSectionSlice","name":"EbookPostsSectionSlice","slice_type":"posts_section_slice","items":[{"theme":"light","data":{"link_type":"Web","url":"http://twitter.com"}},{"theme":"light","data":{"link_type":"Web","url":"http://twitter.com"}},{"theme":"light","data":{"link_type":"Web","url":"https://prismic.io"}}],"primary":{"animation":"fade-right","title":[{"type":"heading1","text":"Strategize turn-key networks","spans":[]}],"containerSize":"1026"},"id":"_EbookPostsSectionSlice"}
     }
   },
   template: '<SliceZone :slices="[mock]" :resolver="resolve" />'
