+++ client/prismicSlices/pageParts/CardsWithIcon/model.json
@@ -10,7 +10,43 @@
       "docURL": "...",
       "version": "sktwi1xtmkfgx8626",
       "description": "CardsWithIcon",
-      "primary": {},
+      "primary": {
+        "animation": {
+          "config": {
+            "label": "Animation List",
+            "placeholder": "",
+            "options": [
+              "fade-up",
+              "fade-down",
+              "fade-right",
+              "fade-left",
+              "fade-up-right",
+              "fade-up-left",
+              "fade-down-right",
+              "fade-down-left",
+              "zoom-in",
+              "zoom-in-down",
+              "zoom-in-left",
+              "zoom-in-right",
+              "none"
+            ],
+            "default_value": "fade-up"
+          },
+          "type": "Select"
+        },
+        "containerSize": {
+          "config": {
+            "label": "Container Size",
+            "placeholder": "",
+            "options": [
+              "1026",
+              "1240"
+            ],
+            "default_value": "1026"
+          },
+          "type": "Select"
+        }
+      },
       "items": {
         "iconImage": {
           "config": {
